+++ src/Conduit.Mediator/INotification.cs
@@ -1,9 +1,0 @@
-namespace Conduit.Mediator;
-
-/// <summary>
-/// Marker interface to represent a notification.
-/// Notifications are published to multiple handlers and do not return a response.
-/// </summary>
-public interface INotification
-{
-}