+++ src/Conduit.Mediator/IStreamRequest.cs
@@ -1,9 +1,0 @@
-namespace Conduit.Mediator;
-
-/// <summary>
-/// Marker interface to represent a request that streams responses.
-/// </summary>
-/// <typeparam name="TResponse">Stream item type</typeparam>
-public interface IStreamRequest<out TResponse>
-{
-}